src/backend/requirements.txt
fastapi
uvicorn
autogen-agentchat
autogen-ext[azure]
autogen-ext[openai]
azure-cosmos
azure-identity
azure-storage-blob
python-dotenv
python-multipart
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-openai
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
reportlab
aiohttp
numpy
pandas
yfinance
tenacity
langchain
sec_api
mplfinance
azure-monitor-opentelemetry
azure-monitor-events-extension
ta